refactor(message): isolate markdown stream cache
