+++ scripts/run-tests.mjs
@@ -315,6 +315,7 @@ export const SUITES = {
     "src/components/chat-thread-rail.test.ts",
     "src/components/chat-rail-modern-redesign.test.ts",
     "src/components/message-bubble-markdown.test.ts",
+    "src/lib/message-markdown-stream.test.ts",
     "src/lib/markdown-preview-shell.test.ts",
     "src/lib/markdown-table-cells.test.ts",
     "src/components/mermaid-viewer.test.ts",
